+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Added Logit Boost classifier
 
 - 1.1.2
+    - Deduplicate Preset seeder centroids
     - Fix Gradient Boost learning rate upper bound
 
 - 1.1.1